When editing tags, exclude inherited tags
Solves the problem that when adding and removing tags, all the file
tags where also re-inserted at the specific node. This basically made
tag inheritance less useful since one couldn't "switch off" tags by
merely removing them from the file tags line in the header -- it could
be the case, in the previous version, that they are repeated in a
local node. And this, in turn, could happen when editing the node
remotely, that is, without the user ever becoming aware of it.