Fix tests for backend inheritance
The previous version raised an error since the expected contents of
the slot :printers was in the wrong order. Actually the order is of no
importance, the list in :printers being an alist which is accessed by
its keys, so we just adapt it to whatever result the current merging
algorithm returns.