delve--ask-storage-file-name: rewrite using the new paths scheme
We prompt the user for a file name defaulting to DELVE--STORAGE-DIR,
which is the first value in DELVE-STORAGE-PATHS. If a new name has
been entered successfully, set DELVE--STORAGE-DIR to its directory, so
that new invocations of DELVE--ASK-STORAGE-FILE-NAME remember that
directory location.

File name completion has been wrapped in a while loop to give some
user feedback for invalid names.