Handle aliased links when checking DB retrieval
delve-query-nodes-by-id: After querying the nodes, only inform the
user about a mismatch of requested IDs and the obtained result if it
is not due to aliased nodes which share the same ID, thereby reducing
the result set without actually indicating that the DB is out of sync.

Also demote the user-error which informs about a potential de-sync to
a mere message.