Use alternative function for completing-read-multiple
The built-in crm does not work with long names, so we had relied on
consult.el. Now their crm function has been removed, so we use an
alternative which is going to be used in citar.el, it seems. See
delve-crm.el for the link to the PR and its authors.