Fix documentation on initializing Delve with use-package
I had always had the problem that when I open a note the first time
with Org-Roam, Delve still did not get loaded. Adding the keyword
`:demand t` ensures that this package will be loaded directly after
Org-Roam, as it is specified in the use-package keyword  `:after`.

See https://github.com/jwiegley/use-package#loading-packages-in-sequence